+++ frugalos_segment/src/lib.rs
@@ -32,7 +32,7 @@ extern crate slog;
 #[macro_use]
 extern crate trackable;
 
-pub use client::storage::{build_ec, ErasureCoder};
+pub use client::ec::{build_ec, ErasureCoder};
 pub use client::Client;
 pub use error::{Error, ErrorKind};
 pub use service::{Service, ServiceHandle};
